Implemented process management functions
Implemented mdb_init() and mdb_close(), reversed opipe and ipipe
to be from the child process's point of view instead of the parent
processes, and fixed a couple misc warnings.
I may want to add a way to force the child process to terminate
if it doesn't terminate within a given timeout period, but for now
I'm using a standard wait() sys call to wait for it to terminate.