Fixed  mdb_step() bug; added MDB_ERR() and MDB_DBG()
The function itself was not bugged, but rather it was called
incorrectly.
Also defined macros for printing error and debug information.
Probably should improve these in the future to be more flexable,
and probably should relocate from mdblib.h to mdblib.c, due to not
being a part of the public interface.