Fixed mdb_write_mem()
Finally figured out why I was failing to read back the value I
thought I was writing to the Pic. The loop that compiles all of
the `mdbword` elements from `wordv[]` into a c string looped up to
`size` (defined as the amount of memory needed to store said c
string) rather than `wordc`, which is the user-provided length of
`wordv`.