Break and watch functions now return breakpoint number
Each break and watch function now extract and return breakpoint
numbers from the output of each of the mdb commands instead of, as
previously implemented, trying to parse the output of "info break"
for the breakpoint numbers.
Also implemented mdb_state() to return the current state of the
simulation being debugged and mdb_noop() so that the running state
can be updated in the event of a breakpoint without having to pass
an actual command.