Fixed mdb_quit()
Previously this would cause an erroneous situation where the
output of mdb would look like a breakpoint would hit, and a re-
read would be attempted. Not a major issue since this would
usually only happen on quitting out of execution anyway, but it
would prevent a clean exit. The solution was to find "quit" in
mdb output since the output includes the command passed.
Note: technically the error and solution are in `mdb_get()`, but
the error only arose as a result of calling `mdb_quit()`.